Replace deprecated (double) casts with (float)
PHP 8.5 deprecates the non-canonical (double) cast. Because one of the
eight occurrences sits in AbstractWrapper::cast(), every parsed response
that contains a float-typed field raises E_DEPRECATED — on a WordPress
site with WP_DEBUG enabled that fills the debug log on every API call.

(float) is the canonical spelling of the same cast, so behaviour is
unchanged.

+++ src/Justimmo/Model/Wrapper/V1/AbstractWrapper.php
@@ -51,7 +51,7 @@ abstract class AbstractWrapper implements WrapperInterface
             case 'int':
                 return (int) $xml;
             case 'double':
-                return (double) $xml;
+                return (float) $xml;
             case 'boolean' :
                 return (bool) ((string) $xml);
             case 'datetime':
